requirements.txt
aiogram==3.*
fastapi==0.110.*
uvicorn[standard]==0.29.*
aiohttp==3.*
loguru==0.7.*
pydantic==2.*
python-dotenv==1.*
redis==5.*
prometheus-fastapi-instrumentator==7.*
PyJWT==2.*
openai==1.*
pygsheets
pytest==7.*
pytest-asyncio==0.21.*